server/ai_service/requirements.txt
sentence-transformers
scikit-learn
psycopg2-binary
web3
python-dotenv
pandas
